refactor(sandbox-fc): extract network pool firewall
